+++ assets/src/apps/authoring/store/groups/layouts/deck/actions/updateActivityPartInheritance.ts
@@ -1,4 +1,5 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
+import { ActivityUpdate, edit } from 'data/persistence/activity';
 import {
   selectActivityById,
   upsertActivities,
@@ -8,6 +9,10 @@ import {
   DeckLayoutGroup,
   GroupsSlice,
 } from '../../../../../../delivery/store/features/groups/slice';
+import { acquireEditingLock, releaseEditingLock } from '../../../../app/actions/locking';
+import { selectProjectSlug } from '../../../../app/slice';
+import { selectResourceId } from '../../../../page/slice';
+import {isEqual} from 'lodash'
 
 export const updateActivityPartInheritance = createAsyncThunk(
   `${GroupsSlice}/updateActivityPartInheritance`,
@@ -48,16 +53,32 @@ export const updateActivityPartInheritance = createAsyncThunk(
       if (!childActivity) {
         return;
       }
-      if (JSON.stringify(childActivity.model.authoring.parts) !== JSON.stringify(combinedParts)) {
+
+      if (!isEqual(childActivity.model.authoring.parts, combinedParts)) {
         const clone = JSON.parse(JSON.stringify(childActivity));
         clone.model.authoring.parts = combinedParts;
         activitiesToUpdate.push(clone);
       }
     });
     if (activitiesToUpdate.length) {
-      console.log('UPDATE: ', { activitiesToUpdate });
+      await dispatch(acquireEditingLock());
+      /* console.log('UPDATE: ', { activitiesToUpdate }); */
       dispatch(upsertActivities({ activities: activitiesToUpdate }));
       // TODO: write to server
+      const projectSlug = selectProjectSlug(rootState);
+      const resourceId = selectResourceId(rootState);
+      // in lieu of bulk edit
+      const updates = activitiesToUpdate.map((activity) => {
+        const changeData: ActivityUpdate = {
+          title: activity.title,
+          objectives: activity.objectives,
+          content: activity.model,
+        };
+        return edit(projectSlug, resourceId, activity.activity_id, changeData, false);
+      });
+      await Promise.all(updates);
+      await dispatch(releaseEditingLock());
+      return;
     }
   },
 );