+++ assets/src/apps/authoring/store/page/slice.ts
@@ -1,4 +1,5 @@
 import { createSelector, createSlice, PayloadAction, Slice } from '@reduxjs/toolkit';
+import { ResourceId } from 'data/types';
 import { RootState } from '../rootReducer';
 
 export interface PageState {
@@ -6,6 +7,8 @@ export interface PageState {
   authorEmail: string;
   objectives: any;
   title: string;
+  revisionSlug: string;
+  resourceId: ResourceId;
 }
 
 const initialState: PageState = {
@@ -13,6 +16,8 @@ const initialState: PageState = {
   authorEmail: '',
   objectives: { attached: [] },
   title: 'New Adaptive Page',
+  revisionSlug: '',
+  resourceId: -1,
 };
 
 const slice: Slice<PageState> = createSlice({
@@ -24,6 +29,8 @@ const slice: Slice<PageState> = createSlice({
       state.authorEmail = action.payload.authorEmail || initialState.authorEmail;
       state.title = action.payload.title || initialState.title;
       state.objectives = action.payload.objectives || initialState.objectives;
+      state.resourceId = action.payload.resourceId || initialState.resourceId;
+      state.revisionSlug = action.payload.revisionSlug || initialState.revisionSlug;
     },
     setIsGraded(state, action: PayloadAction<{ graded: boolean }>) {
       state.graded = action.payload.graded;
@@ -45,5 +52,10 @@ export const selectState = (state: RootState): PageState => state[PageSlice] as 
 export const selectIsGraded = createSelector(selectState, (state: PageState) => state.graded);
 export const selectTitle = createSelector(selectState, (state: PageState) => state.title);
 export const selectObjectives = createSelector(selectState, (state: PageState) => state.objectives);
+export const selectResourceId = createSelector(selectState, (state: PageState) => state.resourceId);
+export const selectRevisionSlug = createSelector(
+  selectState,
+  (state: PageState) => state.revisionSlug,
+);
 
 export default slice.reducer;