+++ assets/src/apps/delivery/layouts/deck/DeckLayoutFooter.tsx
@@ -124,19 +124,19 @@ const DeckLayoutFooter: React.FC = () => {
   }, [lastCheckTimestamp]);
 
   useEffect(() => {
-    if (!lastCheckResults || !lastCheckResults.length) {
+    if (!lastCheckResults || !lastCheckResults.results.length) {
       return;
     }
     // when this changes, notify check has completed
 
-    const isCorrect = lastCheckResults.every((r) => r.params.correct);
+    const isCorrect = lastCheckResults.results.every((r: any) => r.params.correct);
 
     // depending on combineFeedback value is whether we should address more than one event
     const combineFeedback = !!currentActivity.custom.combineFeedback;
 
-    let eventsToProcess = [lastCheckResults[0]];
+    let eventsToProcess = [lastCheckResults.results[0]];
     if (combineFeedback) {
-      eventsToProcess = lastCheckResults;
+      eventsToProcess = lastCheckResults.results;
     }
 
     const actionsByType: any = {
@@ -386,7 +386,6 @@ const DeckLayoutFooter: React.FC = () => {
   }, [currentActivity]);
 
   const currentActivityIds = (currentActivityTree || []).map((a) => a.id);
-
   return (
     <div className={containerClasses.join(' ')} style={{ width: containerWidth }}>
       <NextButton
@@ -431,7 +430,7 @@ const DeckLayoutFooter: React.FC = () => {
               </button>
             </div>
             <style type="text/css" aria-hidden="true" />
-            <div className="content">
+            <div className="content" style={{overflow:'hidden auto !important'}}>
               <FeedbackRenderer
                 feedbacks={currentFeedbacks}
                 snapshot={getLocalizedStateSnapshot(currentActivityIds)}