+++ assets/src/components/activities/DeliveryElement.tsx
@@ -1,4 +1,5 @@
 import { EventEmitter } from 'events';
+import { Maybe } from 'tsmonad';
 import { valueOr } from 'utils/common';
 import {
   Action,
@@ -11,6 +12,8 @@ import {
   StudentResponse,
   Success,
 } from './types';
+import React, { useContext } from 'react';
+import { defaultWriterContext, WriterContext } from 'data/content/writers/context';
 
 export interface EvaluationResponse extends Success {
   actions: Action[];
@@ -234,3 +237,26 @@ export abstract class DeliveryElement<T extends ActivityModelSchema> extends HTM
     }
   }
 }
+
+export interface DeliveryElementState<T> extends DeliveryElementProps<T> {
+  writerContext: WriterContext;
+}
+const DeliveryElementContext = React.createContext<DeliveryElementState<any> | undefined>(
+  undefined,
+);
+export function useDeliveryElementContext<T>() {
+  return Maybe.maybe(
+    useContext<DeliveryElementState<T> | undefined>(DeliveryElementContext),
+  ).valueOrThrow(
+    new Error('useDeliveryElementContext must be used within an DeliveryElementProvider'),
+  );
+}
+export const DeliveryElementProvider: React.FC<DeliveryElementProps<any>> = (props) => {
+  const writerContext = defaultWriterContext({ sectionSlug: props.sectionSlug });
+
+  return (
+    <DeliveryElementContext.Provider value={{ ...props, writerContext }}>
+      {props.children}
+    </DeliveryElementContext.Provider>
+  );
+};