+++ assets/src/components/activities/short_answer/ShortAnswerDelivery.tsx
@@ -12,7 +12,7 @@ import * as ActivityTypes from '../types';
 import { Stem } from '../common/DisplayedStem';
 import { Hints } from '../common/DisplayedHints';
 import { Reset } from '../common/Reset';
-import { Evaluation } from '../common/Evaluation';
+import { Evaluation } from '../common/delivery/evaluation/Evaluation';
 import { valueOr } from 'utils/common';
 import { IconCorrect, IconIncorrect } from 'components/misc/Icons';
 import { defaultWriterContext } from 'data/content/writers/context';