+++ assets/src/data/content/activities/activityUtils.ts
@@ -1,0 +1,3 @@
+import { ActivityState } from 'components/activities/types';
+
+export const isCorrect = (activityState: ActivityState) => activityState.score !== 0;