+++ lib/oli/authoring/editing/page_editor.ex
@@ -333,6 +333,7 @@ defmodule Oli.Authoring.Editing.PageEditor do
       {:ok,
        Publishing.get_published_activity_revisions(publication_id, found_activities)
        |> Enum.map(fn %Revision{
+                        resource_id: activity_id,
                         activity_type_id: activity_type_id,
                         objectives: objectives,
                         slug: slug,
@@ -342,6 +343,7 @@ defmodule Oli.Authoring.Editing.PageEditor do
            type: "activity",
            typeSlug: Map.get(id_to_slug, activity_type_id),
            activitySlug: slug,
+           activity_id: activity_id,
            model: content,
            objectives: objectives
          }