+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## Unreleased
 
+### Bug fixes
+
+- Fix an issue with image coding activity in preview
+
 ### Enhancements
 
 - Add ability to generate and download a course digest from existing course projects
-- Redesign check all that apply activity
+- Redesign check all that apply, multiple choice activities
 - Merge activity editing into the page editor
 
 ## 0.11.1 (2021-6-16)